+ Added package com.github.mafelp.utils ➡️Moved com.github.mafelp.Settings to com.github.mafelp.utils.Settings ➡️Moved com.github.mafelp.Logging to com.github.mafelp.utils.Logging 📗 Updated Javadoc
