+++ doc/allclasses-index.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>All Classes</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="class index">
 <meta name="generator" content="javadoc/AllClassesIndexWriter">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
@@ -71,6 +71,18 @@ loadScripts(document, 'script');</script>
 </th>
 </tr>
 <tr class="row-color" id="i1">
+<td class="col-first"><a href="com/github/mafelp/minecraft/commands/Config.html" title="class in com.github.mafelp.minecraft.commands">Config</a></td>
+<th class="col-last" scope="row">
+<div class="block">Class that implements the /config command in the minecraft chat.</div>
+</th>
+</tr>
+<tr class="alt-color" id="i2">
+<td class="col-first"><a href="com/github/mafelp/discord/commands/CreateChannelListener.html" title="class in com.github.mafelp.discord.commands">CreateChannelListener</a></td>
+<th class="col-last" scope="row">
+<div class="block">The class that listens to the discord chats, if the channel creation command is executed.</div>
+</th>
+</tr>
+<tr class="row-color" id="i3">
 <td class="col-first"><a href="com/github/mafelp/discord/DiscordListener.html" title="class in com.github.mafelp.discord">DiscordListener</a></td>
 <th class="col-last" scope="row">
 <div class="block">Class that listens to all messages sent to a discord server
@@ -78,19 +90,19 @@ loadScripts(document, 'script');</script>
  via direct message.</div>
 </th>
 </tr>
-<tr class="alt-color" id="i2">
+<tr class="alt-color" id="i4">
 <td class="col-first"><a href="com/github/mafelp/discord/DiscordMain.html" title="class in com.github.mafelp.discord">DiscordMain</a></td>
 <th class="col-last" scope="row">
 <div class="block">The class that handles initiation and destruction of the discord bot instance(s)</div>
 </th>
 </tr>
-<tr class="row-color" id="i3">
+<tr class="row-color" id="i5">
 <td class="col-first"><a href="com/github/mafelp/minecraft/JoinListener.html" title="class in com.github.mafelp.minecraft">JoinListener</a></td>
 <th class="col-last" scope="row">
 <div class="block">Class handling Join events of players joining the server</div>
 </th>
 </tr>
-<tr class="alt-color" id="i4">
+<tr class="alt-color" id="i6">
 <td class="col-first"><a href="com/github/mafelp/minecraft/commands/Link.html" title="class in com.github.mafelp.minecraft.commands">Link</a></td>
 <th class="col-last" scope="row">
 <div class="block">Class handling the linking of minecraft and discord accounts -
@@ -98,7 +110,13 @@ loadScripts(document, 'script');</script>
  Warning This class in currently not implemented.</div>
 </th>
 </tr>
-<tr class="row-color" id="i5">
+<tr class="row-color" id="i7">
+<td class="col-first"><a href="com/github/mafelp/utils/Logging.html" title="class in com.github.mafelp.utils">Logging</a></td>
+<th class="col-last" scope="row">
+<div class="block">Logging class to make logging easier and implement a global logging syntax.</div>
+</th>
+</tr>
+<tr class="alt-color" id="i8">
 <td class="col-first"><a href="com/github/mafelp/minecraft/Main.html" title="class in com.github.mafelp.minecraft">Main</a></td>
 <th class="col-last" scope="row">
 <div class="block">The main class which includes enable and
@@ -106,13 +124,13 @@ loadScripts(document, 'script');</script>
  and shutdown or on reload.</div>
 </th>
 </tr>
-<tr class="alt-color" id="i6">
+<tr class="row-color" id="i9">
 <td class="col-first"><a href="com/github/mafelp/minecraft/MinecraftChatListener.html" title="class in com.github.mafelp.minecraft">MinecraftChatListener</a></td>
 <th class="col-last" scope="row">
 <div class="block">Class that reacts to chat messages sent on the minecraft server.</div>
 </th>
 </tr>
-<tr class="row-color" id="i7">
+<tr class="alt-color" id="i10">
 <td class="col-first"><a href="com/github/mafelp/discord/RoleAdmin.html" title="class in com.github.mafelp.discord">RoleAdmin</a></td>
 <th class="col-last" scope="row">
 <div class="block">Class to manage the role that allows users to see the channel,
@@ -121,8 +139,8 @@ loadScripts(document, 'script');</script>
  Warning This class in currently not implemented.</div>
 </th>
 </tr>
-<tr class="alt-color" id="i8">
-<td class="col-first"><a href="com/github/mafelp/Settings.html" title="class in com.github.mafelp">Settings</a></td>
+<tr class="row-color" id="i11">
+<td class="col-first"><a href="com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils">Settings</a></td>
 <th class="col-last" scope="row">
 <div class="block">Class containing all settings:
  [
@@ -132,20 +150,20 @@ loadScripts(document, 'script');</script>
  ]</div>
 </th>
 </tr>
-<tr class="row-color" id="i9">
+<tr class="alt-color" id="i12">
 <td class="col-first"><a href="com/github/mafelp/minecraft/skins/Skin.html" title="class in com.github.mafelp.minecraft.skins">Skin</a></td>
 <th class="col-last" scope="row">
 <div class="block">Class with information about a Player's skin</div>
 </th>
 </tr>
-<tr class="alt-color" id="i10">
+<tr class="row-color" id="i13">
 <td class="col-first"><a href="com/github/mafelp/minecraft/skins/SkinManager.html" title="class in com.github.mafelp.minecraft.skins">SkinManager</a></td>
 <th class="col-last" scope="row">
 <div class="block">Skin Manager used for managing getting skins from Mojang, saving them and getting head
  images for the user avatar in Discord Messages.</div>
 </th>
 </tr>
-<tr class="row-color" id="i11">
+<tr class="alt-color" id="i14">
 <td class="col-first"><a href="com/github/mafelp/minecraft/commands/Token.html" title="class in com.github.mafelp.minecraft.commands">Token</a></td>
 <th class="col-last" scope="row">
 <div class="block">"/token" command - use: set the discord api token with a command,
