+++ doc/allpackages-index.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>All Packages</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="package index">
 <meta name="generator" content="javadoc/AllPackagesIndexWriter">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
@@ -64,11 +64,11 @@ loadScripts(document, 'script');</script>
 </thead>
 <tbody>
 <tr class="alt-color">
-<th class="col-first" scope="row"><a href="com/github/mafelp/package-summary.html">com.github.mafelp</a></th>
+<th class="col-first" scope="row"><a href="com/github/mafelp/discord/package-summary.html">com.github.mafelp.discord</a></th>
 <td class="col-last">&nbsp;</td>
 </tr>
 <tr class="row-color">
-<th class="col-first" scope="row"><a href="com/github/mafelp/discord/package-summary.html">com.github.mafelp.discord</a></th>
+<th class="col-first" scope="row"><a href="com/github/mafelp/discord/commands/package-summary.html">com.github.mafelp.discord.commands</a></th>
 <td class="col-last">&nbsp;</td>
 </tr>
 <tr class="alt-color">
@@ -81,6 +81,10 @@ loadScripts(document, 'script');</script>
 </tr>
 <tr class="alt-color">
 <th class="col-first" scope="row"><a href="com/github/mafelp/minecraft/skins/package-summary.html">com.github.mafelp.minecraft.skins</a></th>
+<td class="col-last">&nbsp;</td>
+</tr>
+<tr class="row-color">
+<th class="col-first" scope="row"><a href="com/github/mafelp/utils/package-summary.html">com.github.mafelp.utils</a></th>
 <td class="col-last">&nbsp;</td>
 </tr>
 </tbody>
