+++ doc/com/github/mafelp/Settings.html
@@ -77,7 +77,7 @@ loadScripts(document, 'script');</script>
 <h1 title="Class Settings" class="title">Class Settings</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree">java.lang.Object
-<div class="inheritance">com.github.mafelp.Settings</div>
+<div class="inheritance">com.github.mafelp.utils.Settings</div>
 </div>
 <section class="description">
 <hr>
