+++ doc/com/github/mafelp/discord/ChannelAdmin.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>ChannelAdmin</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="declaration: package: com.github.mafelp.discord, class: ChannelAdmin">
 <meta name="generator" content="javadoc/ClassWriterImpl">
 <link rel="stylesheet" type="text/css" href="../../../../stylesheet.css" title="Style">
@@ -136,11 +136,13 @@ java.lang.String&nbsp;message)</code></th>
 </td>
 </tr>
 <tr class="row-color" id="i1">
-<td class="col-first"><code>static boolean</code></td>
-<th class="col-second" scope="row"><code><span class="member-name-link"><a href="#createChannel(java.lang.String,org.javacord.api.entity.server.Server,org.javacord.api.entity.channel.ChannelCategory,java.lang.String)">createChannel</a></span>&#8203;(java.lang.String&nbsp;name,
+<td class="col-first"><code>static org.javacord.api.entity.channel.ServerTextChannel</code></td>
+<th class="col-second" scope="row"><code><span class="member-name-link"><a href="#createChannel(java.lang.String,org.javacord.api.entity.server.Server,java.lang.String,org.javacord.api.entity.message.embed.EmbedBuilder,org.javacord.api.entity.channel.ServerTextChannel,org.javacord.api.entity.message.embed.EmbedBuilder)">createChannel</a></span>&#8203;(java.lang.String&nbsp;name,
 org.javacord.api.entity.server.Server&nbsp;server,
-org.javacord.api.entity.channel.ChannelCategory&nbsp;category,
-java.lang.String&nbsp;topic)</code></th>
+java.lang.String&nbsp;topic,
+org.javacord.api.entity.message.embed.EmbedBuilder&nbsp;successEmbed,
+org.javacord.api.entity.channel.ServerTextChannel&nbsp;successChannel,
+org.javacord.api.entity.message.embed.EmbedBuilder&nbsp;welcomeEmbed)</code></th>
 <td class="col-last">
 <div class="block">Creates a channel to send the minecraft messages to.</div>
 </td>
@@ -178,24 +180,25 @@ java.lang.String&nbsp;topic)</code></th>
 <h2>Method Details</h2>
 <ul class="member-list">
 <li>
-<section class="detail" id="createChannel(java.lang.String,org.javacord.api.entity.server.Server,org.javacord.api.entity.channel.ChannelCategory,java.lang.String)">
+<section class="detail" id="createChannel(java.lang.String,org.javacord.api.entity.server.Server,java.lang.String,org.javacord.api.entity.message.embed.EmbedBuilder,org.javacord.api.entity.channel.ServerTextChannel,org.javacord.api.entity.message.embed.EmbedBuilder)">
 <h3>createChannel</h3>
-<div class="member-signature"><span class="modifiers">public static</span>&nbsp;<span class="return-type">boolean</span>&nbsp;<span class="member-name">createChannel</span>&#8203;(<span class="parameters">java.lang.String&nbsp;name,
+<div class="member-signature"><span class="modifiers">public static</span>&nbsp;<span class="return-type">org.javacord.api.entity.channel.ServerTextChannel</span>&nbsp;<span class="member-name">createChannel</span>&#8203;(<span class="parameters">java.lang.String&nbsp;name,
 org.javacord.api.entity.server.Server&nbsp;server,
-org.javacord.api.entity.channel.ChannelCategory&nbsp;category,
-java.lang.String&nbsp;topic)</span>
-                             throws <span class="exceptions">java.lang.InterruptedException</span></div>
+java.lang.String&nbsp;topic,
+org.javacord.api.entity.message.embed.EmbedBuilder&nbsp;successEmbed,
+org.javacord.api.entity.channel.ServerTextChannel&nbsp;successChannel,
+org.javacord.api.entity.message.embed.EmbedBuilder&nbsp;welcomeEmbed)</span></div>
 <div class="block">Creates a channel to send the minecraft messages to.</div>
 <dl class="notes">
 <dt>Parameters:</dt>
 <dd><code>name</code> - name of the channel</dd>
 <dd><code>server</code> - server on which to create the channel on</dd>
-<dd><code>category</code> - category in which the channel should be created</dd>
 <dd><code>topic</code> - the topic the channel should have</dd>
+<dd><code>successEmbed</code> - the embed to be sent into successChannel after completion</dd>
+<dd><code>successChannel</code> - the channel successEmbed is sent to</dd>
+<dd><code>welcomeEmbed</code> - the embed to sent to the newly created channel</dd>
 <dt>Returns:</dt>
-<dd>success state</dd>
-<dt>Throws:</dt>
-<dd><code>java.lang.InterruptedException</code> - thrown when the thread is being stopped before the waiting loop could be fulfilled</dd>
+<dd>the newly created channel</dd>
 </dl>
 </section>
 </li>
