+++ doc/com/github/mafelp/minecraft/commands/package-summary.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>com.github.mafelp.minecraft.commands</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="declaration: package: com.github.mafelp.minecraft.commands">
 <meta name="generator" content="javadoc/PackageWriterImpl">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
@@ -67,6 +67,12 @@ loadScripts(document, 'script');</script>
 </thead>
 <tbody>
 <tr class="alt-color">
+<th class="col-first" scope="row"><a href="Config.html" title="class in com.github.mafelp.minecraft.commands">Config</a></th>
+<td class="col-last">
+<div class="block">Class that implements the /config command in the minecraft chat.</div>
+</td>
+</tr>
+<tr class="row-color">
 <th class="col-first" scope="row"><a href="Link.html" title="class in com.github.mafelp.minecraft.commands">Link</a></th>
 <td class="col-last">
 <div class="block">Class handling the linking of minecraft and discord accounts -
@@ -74,7 +80,7 @@ loadScripts(document, 'script');</script>
  Warning This class in currently not implemented.</div>
 </td>
 </tr>
-<tr class="row-color">
+<tr class="alt-color">
 <th class="col-first" scope="row"><a href="Token.html" title="class in com.github.mafelp.minecraft.commands">Token</a></th>
 <td class="col-last">
 <div class="block">"/token" command - use: set the discord api token with a command,
