+++ doc/com/github/mafelp/minecraft/skins/package-tree.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>com.github.mafelp.minecraft.skins Class Hierarchy</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="tree: package: com.github.mafelp.minecraft.skins">
 <meta name="generator" content="javadoc/PackageTreeWriter">
 <link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
