+++ doc/constant-values.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>Constant Field Values</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="summary of constants">
 <meta name="generator" content="javadoc/ConstantsSummaryWriterImpl">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
@@ -65,7 +65,7 @@ loadScripts(document, 'script');</script>
 <li>
 <div class="constants-summary">
 <table class="summary-table">
-<caption><span>com.github.mafelp.<a href="com/github/mafelp/Settings.html" title="class in com.github.mafelp">Settings</a></span></caption>
+<caption><span>com.github.mafelp.utils.<a href="com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils">Settings</a></span></caption>
 <thead>
 <tr>
 <th class="col-first" scope="col">Modifier and Type</th>
@@ -75,9 +75,9 @@ loadScripts(document, 'script');</script>
 </thead>
 <tbody>
 <tr class="alt-color">
-<td class="col-first"><code id="com.github.mafelp.Settings.version">public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
-<th class="col-second" scope="row"><code><a href="com/github/mafelp/Settings.html#version">version</a></code></th>
-<td class="col-last"><code>"v0.4.1-beta"</code></td>
+<td class="col-first"><code id="com.github.mafelp.utils.Settings.version">public&nbsp;static&nbsp;final&nbsp;java.lang.String</code></td>
+<th class="col-second" scope="row"><code><a href="com/github/mafelp/utils/Settings.html#version">version</a></code></th>
+<td class="col-last"><code>"v0.6.1-beta"</code></td>
 </tr>
 </tbody>
 </table>
