+++ doc/element-list
@@ -1,5 +1,6 @@
-com.github.mafelp
 com.github.mafelp.discord
+com.github.mafelp.discord.commands
 com.github.mafelp.minecraft
 com.github.mafelp.minecraft.commands
 com.github.mafelp.minecraft.skins
+com.github.mafelp.utils
