+++ doc/index-files/index-12.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>S-Index</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="index: S">
 <meta name="generator" content="javadoc/SplitIndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
@@ -56,17 +56,17 @@ loadScripts(document, 'script');</script>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">I</a>&nbsp;<a href="index-6.html">J</a>&nbsp;<a href="index-7.html">L</a>&nbsp;<a href="index-8.html">M</a>&nbsp;<a href="index-9.html">O</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">V</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
 <h2 class="title" id="I:S">S</h2>
 <dl class="index">
-<dt><span class="member-name-link"><a href="../com/github/mafelp/Settings.html#saveConfiguration()">saveConfiguration()</a></span> - Static method in class com.github.mafelp.<a href="../com/github/mafelp/Settings.html" title="class in com.github.mafelp">Settings</a></dt>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Settings.html#saveConfiguration()">saveConfiguration()</a></span> - Static method in class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils">Settings</a></dt>
 <dd>
 <div class="block">saves the configuration (values) in "YamlConfiguration configuration"
  to the configuration file</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/github/mafelp/Settings.html#serverName">serverName</a></span> - Static variable in class com.github.mafelp.<a href="../com/github/mafelp/Settings.html" title="class in com.github.mafelp">Settings</a></dt>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Settings.html#serverName">serverName</a></span> - Static variable in class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils">Settings</a></dt>
 <dd>
 <div class="block">Name of the server displayed in footer of discord messages
  defined in config.yml</div>
 </dd>
-<dt><a href="../com/github/mafelp/Settings.html" title="class in com.github.mafelp"><span class="type-name-link">Settings</span></a> - Class in <a href="../com/github/mafelp/package-summary.html">com.github.mafelp</a></dt>
+<dt><a href="../com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils"><span class="type-name-link">Settings</span></a> - Class in <a href="../com/github/mafelp/utils/package-summary.html">com.github.mafelp.utils</a></dt>
 <dd>
 <div class="block">Class containing all settings:
  [
@@ -75,9 +75,9 @@ loadScripts(document, 'script');</script>
    handling configuration files
  ]</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/github/mafelp/Settings.html#%3Cinit%3E()">Settings()</a></span> - Constructor for class com.github.mafelp.<a href="../com/github/mafelp/Settings.html" title="class in com.github.mafelp">Settings</a></dt>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Settings.html#%3Cinit%3E()">Settings()</a></span> - Constructor for class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils">Settings</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="member-name-link"><a href="../com/github/mafelp/Settings.html#shortMsg">shortMsg</a></span> - Static variable in class com.github.mafelp.<a href="../com/github/mafelp/Settings.html" title="class in com.github.mafelp">Settings</a></dt>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Settings.html#shortMsg">shortMsg</a></span> - Static variable in class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils">Settings</a></dt>
 <dd>
 <div class="block">Use if message prefix should be one (true) or two (true) lines -
  optionally definable in config.yml -
