+++ doc/index-files/index-2.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>C-Index</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="index: C">
 <meta name="generator" content="javadoc/SplitIndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
@@ -63,9 +63,9 @@ loadScripts(document, 'script');</script>
 </dd>
 <dt><span class="member-name-link"><a href="../com/github/mafelp/discord/ChannelAdmin.html#%3Cinit%3E()">ChannelAdmin()</a></span> - Constructor for class com.github.mafelp.discord.<a href="../com/github/mafelp/discord/ChannelAdmin.html" title="class in com.github.mafelp.discord">ChannelAdmin</a></dt>
 <dd>&nbsp;</dd>
-<dt><a href="../com/github/mafelp/package-summary.html">com.github.mafelp</a> - package com.github.mafelp</dt>
-<dd>&nbsp;</dd>
 <dt><a href="../com/github/mafelp/discord/package-summary.html">com.github.mafelp.discord</a> - package com.github.mafelp.discord</dt>
+<dd>&nbsp;</dd>
+<dt><a href="../com/github/mafelp/discord/commands/package-summary.html">com.github.mafelp.discord.commands</a> - package com.github.mafelp.discord.commands</dt>
 <dd>&nbsp;</dd>
 <dt><a href="../com/github/mafelp/minecraft/package-summary.html">com.github.mafelp.minecraft</a> - package com.github.mafelp.minecraft</dt>
 <dd>&nbsp;</dd>
@@ -73,9 +73,28 @@ loadScripts(document, 'script');</script>
 <dd>&nbsp;</dd>
 <dt><a href="../com/github/mafelp/minecraft/skins/package-summary.html">com.github.mafelp.minecraft.skins</a> - package com.github.mafelp.minecraft.skins</dt>
 <dd>&nbsp;</dd>
-<dt><span class="member-name-link"><a href="../com/github/mafelp/discord/ChannelAdmin.html#createChannel(java.lang.String,org.javacord.api.entity.server.Server,org.javacord.api.entity.channel.ChannelCategory,java.lang.String)">createChannel(String, Server, ChannelCategory, String)</a></span> - Static method in class com.github.mafelp.discord.<a href="../com/github/mafelp/discord/ChannelAdmin.html" title="class in com.github.mafelp.discord">ChannelAdmin</a></dt>
+<dt><a href="../com/github/mafelp/utils/package-summary.html">com.github.mafelp.utils</a> - package com.github.mafelp.utils</dt>
+<dd>&nbsp;</dd>
+<dt><a href="../com/github/mafelp/minecraft/commands/Config.html" title="class in com.github.mafelp.minecraft.commands"><span class="type-name-link">Config</span></a> - Class in <a href="../com/github/mafelp/minecraft/commands/package-summary.html">com.github.mafelp.minecraft.commands</a></dt>
+<dd>
+<div class="block">Class that implements the /config command in the minecraft chat.</div>
+</dd>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/minecraft/commands/Config.html#%3Cinit%3E()">Config()</a></span> - Constructor for class com.github.mafelp.minecraft.commands.<a href="../com/github/mafelp/minecraft/commands/Config.html" title="class in com.github.mafelp.minecraft.commands">Config</a></dt>
+<dd>&nbsp;</dd>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/discord/ChannelAdmin.html#createChannel(java.lang.String,org.javacord.api.entity.server.Server,java.lang.String,org.javacord.api.entity.message.embed.EmbedBuilder,org.javacord.api.entity.channel.ServerTextChannel,org.javacord.api.entity.message.embed.EmbedBuilder)">createChannel(String, Server, String, EmbedBuilder, ServerTextChannel, EmbedBuilder)</a></span> - Static method in class com.github.mafelp.discord.<a href="../com/github/mafelp/discord/ChannelAdmin.html" title="class in com.github.mafelp.discord">ChannelAdmin</a></dt>
 <dd>
 <div class="block">Creates a channel to send the minecraft messages to.</div>
+</dd>
+<dt><a href="../com/github/mafelp/discord/commands/CreateChannelListener.html" title="class in com.github.mafelp.discord.commands"><span class="type-name-link">CreateChannelListener</span></a> - Class in <a href="../com/github/mafelp/discord/commands/package-summary.html">com.github.mafelp.discord.commands</a></dt>
+<dd>
+<div class="block">The class that listens to the discord chats, if the channel creation command is executed.</div>
+</dd>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/discord/commands/CreateChannelListener.html#%3Cinit%3E()">CreateChannelListener()</a></span> - Constructor for class com.github.mafelp.discord.commands.<a href="../com/github/mafelp/discord/commands/CreateChannelListener.html" title="class in com.github.mafelp.discord.commands">CreateChannelListener</a></dt>
+<dd>&nbsp;</dd>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Settings.html#createDefaultConfig()">createDefaultConfig()</a></span> - Static method in class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils">Settings</a></dt>
+<dd>
+<div class="block">creates a default and template configuration to set as default and use
+ when user didn't specify a value in config.yml</div>
 </dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">I</a>&nbsp;<a href="index-6.html">J</a>&nbsp;<a href="index-7.html">L</a>&nbsp;<a href="index-8.html">M</a>&nbsp;<a href="index-9.html">O</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">V</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a></main>
