+++ doc/index-files/index-5.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>I-Index</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="index: I">
 <meta name="generator" content="javadoc/SplitIndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
@@ -56,11 +56,15 @@ loadScripts(document, 'script');</script>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">I</a>&nbsp;<a href="index-6.html">J</a>&nbsp;<a href="index-7.html">L</a>&nbsp;<a href="index-8.html">M</a>&nbsp;<a href="index-9.html">O</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">V</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a>
 <h2 class="title" id="I:I">I</h2>
 <dl class="index">
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Logging.html#info(java.lang.String)">info(String)</a></span> - Static method in class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Logging.html" title="class in com.github.mafelp.utils">Logging</a></dt>
+<dd>
+<div class="block">Log a simple info message with the plugin syntax.</div>
+</dd>
 <dt><span class="member-name-link"><a href="../com/github/mafelp/discord/DiscordMain.html#init()">init()</a></span> - Static method in class com.github.mafelp.discord.<a href="../com/github/mafelp/discord/DiscordMain.html" title="class in com.github.mafelp.discord">DiscordMain</a></dt>
 <dd>
 <div class="block">Method used to create the bot instance and log it in</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/github/mafelp/Settings.html#init()">init()</a></span> - Static method in class com.github.mafelp.<a href="../com/github/mafelp/Settings.html" title="class in com.github.mafelp">Settings</a></dt>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Settings.html#init()">init()</a></span> - Static method in class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils">Settings</a></dt>
 <dd>
 <div class="block">Initializing method of the settings -
  initializes config and variables</div>
