+++ doc/index-files/index-7.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>L-Index</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="index: L">
 <meta name="generator" content="javadoc/SplitIndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
@@ -64,6 +64,24 @@ loadScripts(document, 'script');</script>
 </dd>
 <dt><span class="member-name-link"><a href="../com/github/mafelp/minecraft/commands/Link.html#%3Cinit%3E()">Link()</a></span> - Constructor for class com.github.mafelp.minecraft.commands.<a href="../com/github/mafelp/minecraft/commands/Link.html" title="class in com.github.mafelp.minecraft.commands">Link</a></dt>
 <dd>&nbsp;</dd>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Logging.html#logException(java.lang.Exception,java.lang.String)">logException(Exception, String)</a></span> - Static method in class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Logging.html" title="class in com.github.mafelp.utils">Logging</a></dt>
+<dd>
+<div class="block">Logs a general exception with the general syntax.</div>
+</dd>
+<dt><a href="../com/github/mafelp/utils/Logging.html" title="class in com.github.mafelp.utils"><span class="type-name-link">Logging</span></a> - Class in <a href="../com/github/mafelp/utils/package-summary.html">com.github.mafelp.utils</a></dt>
+<dd>
+<div class="block">Logging class to make logging easier and implement a global logging syntax.</div>
+</dd>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Logging.html#%3Cinit%3E()">Logging()</a></span> - Constructor for class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Logging.html" title="class in com.github.mafelp.utils">Logging</a></dt>
+<dd>&nbsp;</dd>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Logging.html#logInvalidConfigurationException(org.bukkit.configuration.InvalidConfigurationException,java.lang.String)">logInvalidConfigurationException(InvalidConfigurationException, String)</a></span> - Static method in class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Logging.html" title="class in com.github.mafelp.utils">Logging</a></dt>
+<dd>
+<div class="block">Logs an InvalidConfigurationException for YAML configurations.</div>
+</dd>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Logging.html#logIOException(java.io.IOException,java.lang.String)">logIOException(IOException, String)</a></span> - Static method in class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Logging.html" title="class in com.github.mafelp.utils">Logging</a></dt>
+<dd>
+<div class="block">Logs an IOException with the general syntax.</div>
+</dd>
 </dl>
 <a href="index-1.html">B</a>&nbsp;<a href="index-2.html">C</a>&nbsp;<a href="index-3.html">D</a>&nbsp;<a href="index-4.html">G</a>&nbsp;<a href="index-5.html">I</a>&nbsp;<a href="index-6.html">J</a>&nbsp;<a href="index-7.html">L</a>&nbsp;<a href="index-8.html">M</a>&nbsp;<a href="index-9.html">O</a>&nbsp;<a href="index-10.html">P</a>&nbsp;<a href="index-11.html">R</a>&nbsp;<a href="index-12.html">S</a>&nbsp;<a href="index-13.html">T</a>&nbsp;<a href="index-14.html">V</a>&nbsp;<br><a href="../allclasses-index.html">All&nbsp;Classes</a><span class="vertical-separator">|</span><a href="../allpackages-index.html">All&nbsp;Packages</a></main>
 <footer role="contentinfo">
