+++ doc/index-files/index-8.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>M-Index</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="index: M">
 <meta name="generator" content="javadoc/SplitIndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
@@ -70,11 +70,11 @@ loadScripts(document, 'script');</script>
 </dd>
 <dt><span class="member-name-link"><a href="../com/github/mafelp/minecraft/MinecraftChatListener.html#%3Cinit%3E()">MinecraftChatListener()</a></span> - Constructor for class com.github.mafelp.minecraft.<a href="../com/github/mafelp/minecraft/MinecraftChatListener.html" title="class in com.github.mafelp.minecraft">MinecraftChatListener</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="member-name-link"><a href="../com/github/mafelp/Settings.html#minecraftServer">minecraftServer</a></span> - Static variable in class com.github.mafelp.<a href="../com/github/mafelp/Settings.html" title="class in com.github.mafelp">Settings</a></dt>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Settings.html#minecraftServer">minecraftServer</a></span> - Static variable in class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils">Settings</a></dt>
 <dd>
 <div class="block">this minecraft server, defined in Main.onEnable()</div>
 </dd>
-<dt><span class="member-name-link"><a href="../com/github/mafelp/Settings.html#msgPrefix(org.javacord.api.event.message.MessageCreateEvent)">msgPrefix(MessageCreateEvent)</a></span> - Static method in class com.github.mafelp.<a href="../com/github/mafelp/Settings.html" title="class in com.github.mafelp">Settings</a></dt>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/utils/Settings.html#msgPrefix(org.javacord.api.event.message.MessageCreateEvent)">msgPrefix(MessageCreateEvent)</a></span> - Static method in class com.github.mafelp.utils.<a href="../com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils">Settings</a></dt>
 <dd>
 <div class="block">Creates a prefix for a minecraft message</div>
 </dd>
