+++ doc/index-files/index-9.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>O-Index</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="index: O">
 <meta name="generator" content="javadoc/SplitIndexWriter">
 <link rel="stylesheet" type="text/css" href="../stylesheet.css" title="Style">
@@ -60,6 +60,10 @@ loadScripts(document, 'script');</script>
 <dd>
 <div class="block">Method called by the minecraft server when a chat message is being sent</div>
 </dd>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/minecraft/commands/Config.html#onCommand(org.bukkit.command.CommandSender,org.bukkit.command.Command,java.lang.String,java.lang.String%5B%5D)">onCommand(CommandSender, Command, String, String[])</a></span> - Method in class com.github.mafelp.minecraft.commands.<a href="../com/github/mafelp/minecraft/commands/Config.html" title="class in com.github.mafelp.minecraft.commands">Config</a></dt>
+<dd>
+<div class="block">The command executing task.</div>
+</dd>
 <dt><span class="member-name-link"><a href="../com/github/mafelp/minecraft/commands/Link.html#onCommand(org.bukkit.command.CommandSender,org.bukkit.command.Command,java.lang.String,java.lang.String%5B%5D)">onCommand(CommandSender, Command, String, String[])</a></span> - Method in class com.github.mafelp.minecraft.commands.<a href="../com/github/mafelp/minecraft/commands/Link.html" title="class in com.github.mafelp.minecraft.commands">Link</a></dt>
 <dd>
 <div class="block">The Method called when command "/token" is executed.</div>
@@ -79,6 +83,10 @@ loadScripts(document, 'script');</script>
 <dt><span class="member-name-link"><a href="../com/github/mafelp/minecraft/JoinListener.html#onJoin(org.bukkit.event.player.PlayerJoinEvent)">onJoin(PlayerJoinEvent)</a></span> - Method in class com.github.mafelp.minecraft.<a href="../com/github/mafelp/minecraft/JoinListener.html" title="class in com.github.mafelp.minecraft">JoinListener</a></dt>
 <dd>
 <div class="block">Method called by the minecraft server when a player joins.</div>
+</dd>
+<dt><span class="member-name-link"><a href="../com/github/mafelp/discord/commands/CreateChannelListener.html#onMessageCreate(org.javacord.api.event.message.MessageCreateEvent)">onMessageCreate(MessageCreateEvent)</a></span> - Method in class com.github.mafelp.discord.commands.<a href="../com/github/mafelp/discord/commands/CreateChannelListener.html" title="class in com.github.mafelp.discord.commands">CreateChannelListener</a></dt>
+<dd>
+<div class="block">The method called by the discord API, for every chat message.</div>
 </dd>
 <dt><span class="member-name-link"><a href="../com/github/mafelp/discord/DiscordListener.html#onMessageCreate(org.javacord.api.event.message.MessageCreateEvent)">onMessageCreate(MessageCreateEvent)</a></span> - Method in class com.github.mafelp.discord.<a href="../com/github/mafelp/discord/DiscordListener.html" title="class in com.github.mafelp.discord">DiscordListener</a></dt>
 <dd>
