+++ doc/member-search-index.js
@@ -1,1 +1,1 @@
-memberSearchIndex = [{"p":"com.github.mafelp.discord","c":"ChannelAdmin","l":"broadcastMessage(Player, String)","u":"broadcastMessage(org.bukkit.entity.Player,java.lang.String)"},{"p":"com.github.mafelp.discord","c":"ChannelAdmin","l":"ChannelAdmin()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.discord","c":"ChannelAdmin","l":"createChannel(String, Server, ChannelCategory, String)","u":"createChannel(java.lang.String,org.javacord.api.entity.server.Server,org.javacord.api.entity.channel.ChannelCategory,java.lang.String)"},{"p":"com.github.mafelp","c":"Settings","l":"debug"},{"p":"com.github.mafelp","c":"Settings","l":"discordApi"},{"p":"com.github.mafelp.discord","c":"DiscordListener","l":"DiscordListener()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.discord","c":"DiscordMain","l":"DiscordMain()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp","c":"Settings","l":"getApiToken()"},{"p":"com.github.mafelp","c":"Settings","l":"getConfiguration()"},{"p":"com.github.mafelp","c":"Settings","l":"getConfigurationFileDirectory()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getHead()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getHeadFile()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getPlayer()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkin()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkinDirectory()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkinDirectoryHeads()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkinDirectorySkins()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkinFile()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkinFormat()"},{"p":"com.github.mafelp","c":"Settings","l":"init()"},{"p":"com.github.mafelp.discord","c":"DiscordMain","l":"init()"},{"p":"com.github.mafelp.minecraft","c":"JoinListener","l":"JoinListener()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.minecraft.commands","c":"Link","l":"Link()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.minecraft","c":"Main","l":"Main()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.minecraft","c":"MinecraftChatListener","l":"MinecraftChatListener()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp","c":"Settings","l":"minecraftServer"},{"p":"com.github.mafelp","c":"Settings","l":"msgPrefix(MessageCreateEvent)","u":"msgPrefix(org.javacord.api.event.message.MessageCreateEvent)"},{"p":"com.github.mafelp.minecraft","c":"MinecraftChatListener","l":"onChatMessage(AsyncPlayerChatEvent)","u":"onChatMessage(org.bukkit.event.player.AsyncPlayerChatEvent)"},{"p":"com.github.mafelp.minecraft.commands","c":"Link","l":"onCommand(CommandSender, Command, String, String[])","u":"onCommand(org.bukkit.command.CommandSender,org.bukkit.command.Command,java.lang.String,java.lang.String[])"},{"p":"com.github.mafelp.minecraft.commands","c":"Token","l":"onCommand(CommandSender, Command, String, String[])","u":"onCommand(org.bukkit.command.CommandSender,org.bukkit.command.Command,java.lang.String,java.lang.String[])"},{"p":"com.github.mafelp.minecraft","c":"Main","l":"onDisable()"},{"p":"com.github.mafelp.minecraft","c":"Main","l":"onEnable()"},{"p":"com.github.mafelp.minecraft","c":"JoinListener","l":"onJoin(PlayerJoinEvent)","u":"onJoin(org.bukkit.event.player.PlayerJoinEvent)"},{"p":"com.github.mafelp.discord","c":"DiscordListener","l":"onMessageCreate(MessageCreateEvent)","u":"onMessageCreate(org.javacord.api.event.message.MessageCreateEvent)"},{"p":"com.github.mafelp","c":"Settings","l":"prefix"},{"p":"com.github.mafelp.discord","c":"RoleAdmin","l":"RoleAdmin()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp","c":"Settings","l":"saveConfiguration()"},{"p":"com.github.mafelp","c":"Settings","l":"serverName"},{"p":"com.github.mafelp","c":"Settings","l":"Settings()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp","c":"Settings","l":"shortMsg"},{"p":"com.github.mafelp.discord","c":"DiscordMain","l":"shutdown()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"Skin(Player)","u":"%3Cinit%3E(org.bukkit.entity.Player)"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"Skin(Player, boolean)","u":"%3Cinit%3E(org.bukkit.entity.Player,boolean)"},{"p":"com.github.mafelp.minecraft.skins","c":"SkinManager","l":"SkinManager()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.minecraft.commands","c":"Token","l":"Token()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp","c":"Settings","l":"version"}];updateSearchResults();
+memberSearchIndex = [{"p":"com.github.mafelp.discord","c":"ChannelAdmin","l":"broadcastMessage(Player, String)","u":"broadcastMessage(org.bukkit.entity.Player,java.lang.String)"},{"p":"com.github.mafelp.discord","c":"ChannelAdmin","l":"ChannelAdmin()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.minecraft.commands","c":"Config","l":"Config()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.discord","c":"ChannelAdmin","l":"createChannel(String, Server, String, EmbedBuilder, ServerTextChannel, EmbedBuilder)","u":"createChannel(java.lang.String,org.javacord.api.entity.server.Server,java.lang.String,org.javacord.api.entity.message.embed.EmbedBuilder,org.javacord.api.entity.channel.ServerTextChannel,org.javacord.api.entity.message.embed.EmbedBuilder)"},{"p":"com.github.mafelp.discord.commands","c":"CreateChannelListener","l":"CreateChannelListener()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.utils","c":"Settings","l":"createDefaultConfig()"},{"p":"com.github.mafelp.utils","c":"Settings","l":"debug"},{"p":"com.github.mafelp.utils","c":"Settings","l":"discordApi"},{"p":"com.github.mafelp.utils","c":"Settings","l":"discordCommandPrefix"},{"p":"com.github.mafelp.discord","c":"DiscordListener","l":"DiscordListener()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.discord","c":"DiscordMain","l":"DiscordMain()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.utils","c":"Settings","l":"getApiToken()"},{"p":"com.github.mafelp.utils","c":"Settings","l":"getConfiguration()"},{"p":"com.github.mafelp.utils","c":"Settings","l":"getConfigurationFileDirectory()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getHead()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getHeadFile()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getPlayer()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkin()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkinDirectory()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkinDirectoryHeads()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkinDirectorySkins()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkinFile()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"getSkinFormat()"},{"p":"com.github.mafelp.utils","c":"Logging","l":"info(String)","u":"info(java.lang.String)"},{"p":"com.github.mafelp.discord","c":"DiscordMain","l":"init()"},{"p":"com.github.mafelp.utils","c":"Settings","l":"init()"},{"p":"com.github.mafelp.minecraft","c":"JoinListener","l":"JoinListener()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.minecraft.commands","c":"Link","l":"Link()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.utils","c":"Logging","l":"logException(Exception, String)","u":"logException(java.lang.Exception,java.lang.String)"},{"p":"com.github.mafelp.utils","c":"Logging","l":"Logging()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.utils","c":"Logging","l":"logInvalidConfigurationException(InvalidConfigurationException, String)","u":"logInvalidConfigurationException(org.bukkit.configuration.InvalidConfigurationException,java.lang.String)"},{"p":"com.github.mafelp.utils","c":"Logging","l":"logIOException(IOException, String)","u":"logIOException(java.io.IOException,java.lang.String)"},{"p":"com.github.mafelp.minecraft","c":"Main","l":"Main()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.minecraft","c":"MinecraftChatListener","l":"MinecraftChatListener()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.utils","c":"Settings","l":"minecraftServer"},{"p":"com.github.mafelp.utils","c":"Settings","l":"msgPrefix(MessageCreateEvent)","u":"msgPrefix(org.javacord.api.event.message.MessageCreateEvent)"},{"p":"com.github.mafelp.minecraft","c":"MinecraftChatListener","l":"onChatMessage(AsyncPlayerChatEvent)","u":"onChatMessage(org.bukkit.event.player.AsyncPlayerChatEvent)"},{"p":"com.github.mafelp.minecraft.commands","c":"Config","l":"onCommand(CommandSender, Command, String, String[])","u":"onCommand(org.bukkit.command.CommandSender,org.bukkit.command.Command,java.lang.String,java.lang.String[])"},{"p":"com.github.mafelp.minecraft.commands","c":"Link","l":"onCommand(CommandSender, Command, String, String[])","u":"onCommand(org.bukkit.command.CommandSender,org.bukkit.command.Command,java.lang.String,java.lang.String[])"},{"p":"com.github.mafelp.minecraft.commands","c":"Token","l":"onCommand(CommandSender, Command, String, String[])","u":"onCommand(org.bukkit.command.CommandSender,org.bukkit.command.Command,java.lang.String,java.lang.String[])"},{"p":"com.github.mafelp.minecraft","c":"Main","l":"onDisable()"},{"p":"com.github.mafelp.minecraft","c":"Main","l":"onEnable()"},{"p":"com.github.mafelp.minecraft","c":"JoinListener","l":"onJoin(PlayerJoinEvent)","u":"onJoin(org.bukkit.event.player.PlayerJoinEvent)"},{"p":"com.github.mafelp.discord","c":"DiscordListener","l":"onMessageCreate(MessageCreateEvent)","u":"onMessageCreate(org.javacord.api.event.message.MessageCreateEvent)"},{"p":"com.github.mafelp.discord.commands","c":"CreateChannelListener","l":"onMessageCreate(MessageCreateEvent)","u":"onMessageCreate(org.javacord.api.event.message.MessageCreateEvent)"},{"p":"com.github.mafelp.utils","c":"Settings","l":"prefix"},{"p":"com.github.mafelp.discord","c":"RoleAdmin","l":"RoleAdmin()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.utils","c":"Settings","l":"saveConfiguration()"},{"p":"com.github.mafelp.utils","c":"Settings","l":"serverName"},{"p":"com.github.mafelp.utils","c":"Settings","l":"Settings()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.utils","c":"Settings","l":"shortMsg"},{"p":"com.github.mafelp.discord","c":"DiscordMain","l":"shutdown()"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"Skin(Player)","u":"%3Cinit%3E(org.bukkit.entity.Player)"},{"p":"com.github.mafelp.minecraft.skins","c":"Skin","l":"Skin(Player, boolean)","u":"%3Cinit%3E(org.bukkit.entity.Player,boolean)"},{"p":"com.github.mafelp.minecraft.skins","c":"SkinManager","l":"SkinManager()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.minecraft.commands","c":"Token","l":"Token()","u":"%3Cinit%3E()"},{"p":"com.github.mafelp.utils","c":"Settings","l":"version"}];updateSearchResults();
