+++ doc/overview-tree.html
@@ -2,10 +2,10 @@
 <!-- NewPage -->
 <html lang="en">
 <head>
-<!-- Generated by javadoc (15) on Mon Mar 22 19:06:20 CET 2021 -->
+<!-- Generated by javadoc (15) on Thu Mar 25 21:33:18 CET 2021 -->
 <title>Class Hierarchy</title>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
-<meta name="dc.created" content="2021-03-22">
+<meta name="dc.created" content="2021-03-25">
 <meta name="description" content="class tree">
 <meta name="generator" content="javadoc/TreeWriter">
 <link rel="stylesheet" type="text/css" href="stylesheet.css" title="Style">
@@ -54,11 +54,12 @@ loadScripts(document, 'script');</script>
 <h1 class="title">Hierarchy For All Packages</h1>
 <span class="package-hierarchy-label">Package Hierarchies:</span>
 <ul class="horizontal">
-<li><a href="com/github/mafelp/package-tree.html">com.github.mafelp</a>, </li>
 <li><a href="com/github/mafelp/discord/package-tree.html">com.github.mafelp.discord</a>, </li>
+<li><a href="com/github/mafelp/discord/commands/package-tree.html">com.github.mafelp.discord.commands</a>, </li>
 <li><a href="com/github/mafelp/minecraft/package-tree.html">com.github.mafelp.minecraft</a>, </li>
 <li><a href="com/github/mafelp/minecraft/commands/package-tree.html">com.github.mafelp.minecraft.commands</a>, </li>
-<li><a href="com/github/mafelp/minecraft/skins/package-tree.html">com.github.mafelp.minecraft.skins</a></li>
+<li><a href="com/github/mafelp/minecraft/skins/package-tree.html">com.github.mafelp.minecraft.skins</a>, </li>
+<li><a href="com/github/mafelp/utils/package-tree.html">com.github.mafelp.utils</a></li>
 </ul>
 </div>
 <section class="hierarchy">
@@ -67,10 +68,13 @@ loadScripts(document, 'script');</script>
 <li class="circle">java.lang.Object
 <ul>
 <li class="circle">com.github.mafelp.discord.<a href="com/github/mafelp/discord/ChannelAdmin.html" title="class in com.github.mafelp.discord"><span class="type-name-link">ChannelAdmin</span></a></li>
+<li class="circle">com.github.mafelp.minecraft.commands.<a href="com/github/mafelp/minecraft/commands/Config.html" title="class in com.github.mafelp.minecraft.commands"><span class="type-name-link">Config</span></a> (implements org.bukkit.command.CommandExecutor)</li>
+<li class="circle">com.github.mafelp.discord.commands.<a href="com/github/mafelp/discord/commands/CreateChannelListener.html" title="class in com.github.mafelp.discord.commands"><span class="type-name-link">CreateChannelListener</span></a> (implements org.javacord.api.listener.message.MessageCreateListener)</li>
 <li class="circle">com.github.mafelp.discord.<a href="com/github/mafelp/discord/DiscordListener.html" title="class in com.github.mafelp.discord"><span class="type-name-link">DiscordListener</span></a> (implements org.javacord.api.listener.message.MessageCreateListener)</li>
 <li class="circle">com.github.mafelp.discord.<a href="com/github/mafelp/discord/DiscordMain.html" title="class in com.github.mafelp.discord"><span class="type-name-link">DiscordMain</span></a></li>
 <li class="circle">com.github.mafelp.minecraft.<a href="com/github/mafelp/minecraft/JoinListener.html" title="class in com.github.mafelp.minecraft"><span class="type-name-link">JoinListener</span></a> (implements org.bukkit.event.Listener)</li>
 <li class="circle">com.github.mafelp.minecraft.commands.<a href="com/github/mafelp/minecraft/commands/Link.html" title="class in com.github.mafelp.minecraft.commands"><span class="type-name-link">Link</span></a> (implements org.bukkit.command.CommandExecutor)</li>
+<li class="circle">com.github.mafelp.utils.<a href="com/github/mafelp/utils/Logging.html" title="class in com.github.mafelp.utils"><span class="type-name-link">Logging</span></a></li>
 <li class="circle">com.github.mafelp.minecraft.<a href="com/github/mafelp/minecraft/MinecraftChatListener.html" title="class in com.github.mafelp.minecraft"><span class="type-name-link">MinecraftChatListener</span></a> (implements org.bukkit.event.Listener)</li>
 <li class="circle">org.bukkit.plugin.PluginBase (implements org.bukkit.plugin.Plugin)
 <ul>
@@ -82,7 +86,7 @@ loadScripts(document, 'script');</script>
 </ul>
 </li>
 <li class="circle">com.github.mafelp.discord.<a href="com/github/mafelp/discord/RoleAdmin.html" title="class in com.github.mafelp.discord"><span class="type-name-link">RoleAdmin</span></a></li>
-<li class="circle">com.github.mafelp.<a href="com/github/mafelp/Settings.html" title="class in com.github.mafelp"><span class="type-name-link">Settings</span></a></li>
+<li class="circle">com.github.mafelp.utils.<a href="com/github/mafelp/utils/Settings.html" title="class in com.github.mafelp.utils"><span class="type-name-link">Settings</span></a></li>
 <li class="circle">com.github.mafelp.minecraft.skins.<a href="com/github/mafelp/minecraft/skins/Skin.html" title="class in com.github.mafelp.minecraft.skins"><span class="type-name-link">Skin</span></a></li>
 <li class="circle">com.github.mafelp.minecraft.skins.<a href="com/github/mafelp/minecraft/skins/SkinManager.html" title="class in com.github.mafelp.minecraft.skins"><span class="type-name-link">SkinManager</span></a></li>
 <li class="circle">com.github.mafelp.minecraft.commands.<a href="com/github/mafelp/minecraft/commands/Token.html" title="class in com.github.mafelp.minecraft.commands"><span class="type-name-link">Token</span></a> (implements org.bukkit.command.CommandExecutor)</li>
