+++ doc/package-search-index.js
@@ -1,1 +1,1 @@
-packageSearchIndex = [{"l":"All Packages","u":"allpackages-index.html"},{"l":"com.github.mafelp"},{"l":"com.github.mafelp.discord"},{"l":"com.github.mafelp.minecraft"},{"l":"com.github.mafelp.minecraft.commands"},{"l":"com.github.mafelp.minecraft.skins"}];updateSearchResults();
+packageSearchIndex = [{"l":"All Packages","u":"allpackages-index.html"},{"l":"com.github.mafelp.discord"},{"l":"com.github.mafelp.discord.commands"},{"l":"com.github.mafelp.minecraft"},{"l":"com.github.mafelp.minecraft.commands"},{"l":"com.github.mafelp.minecraft.skins"},{"l":"com.github.mafelp.utils"}];updateSearchResults();
