+++ doc/type-search-index.js
@@ -1,1 +1,1 @@
-typeSearchIndex = [{"l":"All Classes","u":"allclasses-index.html"},{"p":"com.github.mafelp.discord","l":"ChannelAdmin"},{"p":"com.github.mafelp.discord","l":"DiscordListener"},{"p":"com.github.mafelp.discord","l":"DiscordMain"},{"p":"com.github.mafelp.minecraft","l":"JoinListener"},{"p":"com.github.mafelp.minecraft.commands","l":"Link"},{"p":"com.github.mafelp.minecraft","l":"Main"},{"p":"com.github.mafelp.minecraft","l":"MinecraftChatListener"},{"p":"com.github.mafelp.discord","l":"RoleAdmin"},{"p":"com.github.mafelp","l":"Settings"},{"p":"com.github.mafelp.minecraft.skins","l":"Skin"},{"p":"com.github.mafelp.minecraft.skins","l":"SkinManager"},{"p":"com.github.mafelp.minecraft.commands","l":"Token"}];updateSearchResults();
+typeSearchIndex = [{"l":"All Classes","u":"allclasses-index.html"},{"p":"com.github.mafelp.discord","l":"ChannelAdmin"},{"p":"com.github.mafelp.minecraft.commands","l":"Config"},{"p":"com.github.mafelp.discord.commands","l":"CreateChannelListener"},{"p":"com.github.mafelp.discord","l":"DiscordListener"},{"p":"com.github.mafelp.discord","l":"DiscordMain"},{"p":"com.github.mafelp.minecraft","l":"JoinListener"},{"p":"com.github.mafelp.minecraft.commands","l":"Link"},{"p":"com.github.mafelp.utils","l":"Logging"},{"p":"com.github.mafelp.minecraft","l":"Main"},{"p":"com.github.mafelp.minecraft","l":"MinecraftChatListener"},{"p":"com.github.mafelp.discord","l":"RoleAdmin"},{"p":"com.github.mafelp.utils","l":"Settings"},{"p":"com.github.mafelp.minecraft.skins","l":"Skin"},{"p":"com.github.mafelp.minecraft.skins","l":"SkinManager"},{"p":"com.github.mafelp.minecraft.commands","l":"Token"}];updateSearchResults();
