+++ src/main/java/com/github/mafelp/utils/Logging.java
@@ -1,16 +1,25 @@
-package com.github.mafelp;
+package com.github.mafelp.utils;
 
 import org.bukkit.configuration.InvalidConfigurationException;
 
 import java.io.IOException;
 
+/**
+ * Logging class to make logging easier and implement a global logging syntax.
+ */
 public class Logging {
+    /**
+     * Logs a general exception with the general syntax.
+     * @param exception The exception that should be logged.
+     * @param logMessage The message that should be logged before the error to describe, where the error occurred.
+     */
     public static void logException(Exception exception, String logMessage) {
-
+        // Log the error
         Settings.minecraftServer.getLogger().warning(
                 Settings.prefix + logMessage + " Error: " + exception.getMessage()
         );
 
+        // Print the Stack trace, when debug mode is enabled.
         if (Settings.debug) {
             for (StackTraceElement s :
                     exception.getStackTrace()) {
@@ -19,11 +28,18 @@ public class Logging {
         }
     }
 
+    /**
+     * Logs an IOException with the general syntax.
+     * @param exception The exception that should be logged.
+     * @param logMessage The message that should be logged before the error to describe, where the error occurred.
+     */
     public static void logIOException(IOException exception, String logMessage) {
+        // Log the error
         Settings.minecraftServer.getLogger().warning(
                 Settings.prefix + logMessage + " Error: " + exception.getMessage()
         );
 
+        // Print the Stack trace, when debug mode is enabled.
         if (Settings.debug) {
             for (StackTraceElement s :
                     exception.getStackTrace()) {
@@ -32,11 +48,19 @@ public class Logging {
         }
     }
 
+    /**
+     * Logs an InvalidConfigurationException for YAML configurations. -
+     * Normally this method will not be called.
+     * @param exception The exception that should be logged.
+     * @param logMessage The message that should be logged before the error to describe, where the error occurred.
+     */
     public static void logInvalidConfigurationException(InvalidConfigurationException exception, String logMessage) {
+        // log the exception
         Settings.minecraftServer.getLogger().warning(
                 Settings.prefix + logMessage + " Error: " + exception.getMessage()
         );
 
+        // Log the stack strace, if debug is enabled.
         if (Settings.debug) {
             for (StackTraceElement stackTraceElement :
                     exception.getStackTrace()) {
@@ -45,6 +69,10 @@ public class Logging {
         }
     }
 
+    /**
+     * Log a simple info message with the plugin syntax.
+     * @param message The message that should be logged.
+     */
     public static void info(String message) {
         Settings.minecraftServer.getLogger().info(Settings.prefix + message);
     }
